Move utilities to own package
