+++ warden/warden.go
@@ -1,18 +1,17 @@
 package warden
 
 import (
-    "net/http"
+	"net/http"
 
-    "github.com/gin-gonic/gin"
+	"github.com/gin-gonic/gin"
 )
 
-
 func allowedHandler(c *gin.Context) {
-    c.JSON(http.StatusOK, gin.H{
-        "allowed": false,
-    })
+	c.JSON(http.StatusOK, gin.H{
+		"allowed": false,
+	})
 }
 
 func SetupRoutes(r *gin.Engine) {
-    r.POST("/allowed", allowedHandler)
+	r.POST("/allowed", allowedHandler)
 }
